--- a/requirements.txt
+++ b/requirements.txt
@@ -7,3 +7,4 @@ scikit-image
 pywavelets
 pyfftw
 dxchange
+numexpr