This PR uses `numexpr` to decrease computation time.

At the moment I applied it to `normalize` only, but I think the same approach could work on some other functions as well. I still have to check thoroughly whether it produces identical output, but the decrease in computation time seems worth it: for a dataset with 1025 projections with 128 slices and 2000 detector pixels, it reduces the time it takes to perform normalization from 1.61 s to 0.249 s on my machine. On another machine it reduces the normalization time from 4.15 s to 0.96 s for a 1025x256x2000 dataset.

I am not really sure what the cause is of the relatively slow performance of the current code on my machines: it could be that the `multiprocessing` module is doing some copying that we are not aware of. This could be related to the remarks at the end of #120.

Maybe the computation time of other functions (such as reconstruction) could be reduced by the same approach. I will investigate this further later, so we can wait with merging this until a later time. 

@dgursoy What are your thoughts on this?