@dgursoy 2016-07-15
This is cool. I haven't used ```numexpr```  before, but this is a good excuse to give it a try. I'll also play with it and share my experiences. The only downside I can tell is that the code looks ugly, but that's fine I guess.