@dmpelt 2016-07-15
It looks a bit ugly indeed, but I think it is still readable, even if someone is not familiar with numexpr. 

We can remove the `local_dict=` things probably. By specifying a local_dict you get slightly better performance since it doesn't have to search the actual larger local dictionary, but in this case the benefits are probably minimal, and the code will look better if we remove it.

What do you think?