@prjemian 2016-07-15
suggest this change:

    if ncore:

to

    if ncore is not None: