@dmpelt 2016-07-15
I normally don't mind `if ncore` like checks for `None`, but an explicit check for `None` is indeed clearer. I could go either way.

I updated the PR to include `minus_log` as well. The previous `minus_log` implementation did not do any multithreading, so the improvements here are even greater: for a 1025x256x2000 dataset it reduces `minus_log` time from 5.232 s to 0.175 s.