@dmpelt 2016-07-21
For tomopy reconstruction algorithms, we don't actually need the `multiprocessing` module, since the `ctypes` module drops the Python GIL, which makes threading with `threading` or `concurrent.futures` possible.

I've just added a commit to this PR that uses `concurrent.futures` instead of `multiprocessing` for reconstruction. On a dataset of size 1025x256x2000, this reduces the computation time for gridrec from around 21 seconds to about 9 seconds on my machine. I think the reason for this is mostly that it avoids copies of the data to `/tmp/` that `multiprocessing` seems to do.

This commit does break ASTRA integration I think: I will try to fix that in a later commit.