@xianghuix 2016-07-21
Daniel,

This sounds very promising. I don't know the details but I heard the hanging issue with gridrec (#120) is related to multiprocessing. I am guessing your change will solve that issue as well? Is it correct?

Thanks,
x
________________________________
From: Daniel M. Pelt [notifications@github.com]
Sent: Wednesday, July 20, 2016 8:05 PM
To: tomopy/tomopy
Subject: Re: [tomopy/tomopy] WIP: Performance improvements (#217)


For tomopy reconstruction algorithms, we don't actually need the multiprocessing module, since the ctypes module drops the Python GIL, which makes threading with threading or concurrent.futures possible.

I've just added a commit to this PR that uses concurrent.futures instead of multiprocessing for reconstruction. On a dataset of size 1025x256x2000, this reduces the computation time for gridrec from around 21 seconds to about 9 seconds on my machine. I think the reason for this is mostly that it avoids copies of the data to /tmp/ that multiprocessing seems to do.

This commit does break ASTRA integration I think: I will try to fix that in a later commit.

—
You are receiving this because you are subscribed to this thread.
Reply to this email directly, view it on GitHub<https://github.com/tomopy/tomopy/pull/217#issuecomment-234129007>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AF0slh9HrvawNiypfoYwDNzEnmbwrR2Rks5qXsXngaJpZM4JM-rM>.