@dmpelt 2016-07-21
Yes, after this commit the gridrec code does not use multiprocessing anymore, so the hanging issue should be fixed. It might now crash instead in these cases, which will at least give us an easier way of diagnosing the underlying problem.