@xianghuix 2016-07-21
That sounds great. Look forward to the new commit.

Best,
x
________________________________
From: Daniel M. Pelt [notifications@github.com]
Sent: Wednesday, July 20, 2016 8:40 PM
To: tomopy/tomopy
Cc: Xiao, Xianghui; Comment
Subject: Re: [tomopy/tomopy] WIP: Performance improvements (#217)


Yes, after this commit the gridrec code does not use multiprocessing anymore, so the hanging issue should be fixed. It might now crash instead in these cases, which will at least give us an easier way of diagnosing the underlying problem.

—
You are receiving this because you commented.
Reply to this email directly, view it on GitHub<https://github.com/tomopy/tomopy/pull/217#issuecomment-234133950>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AF0slpiZgbrYUJ8QKz08Y2mqex2LFsFXks5qXs4XgaJpZM4JM-rM>.