@dmpelt 2016-08-18
I think this can be merged, unless you want to test it a bit more first.

There are a few other routines that I might try to speed up a bit in the future, but I can create new PRs for them if needed.