@dgursoy 2016-08-20
Thanks! Before we forget, could you write few items for the next release notes:
https://github.com/tomopy/tomopy/tree/master/doc/source/release